Initial commit: an ASCII notation for music.
The first draft of a terse, plain-text notation for simple musical
composotions, designed to be "compiled" to MIDI and synthesized with VST
instruments.

In this first commit, an AST and parser are defined with support for
single notes, a selection of conventional chord symbols, rational note
values, a selection of articulation modifiers, and repeats; and a
lowering to multi-track-MIDI files is defined, which unrolls repeats and
expands chord symbols to their constituent notes, and interprets
articulations by adjusting MIDI velocity and note duration.